+++ src/fabscan/server/services/api/FSStreamHandler.py
@@ -8,24 +8,30 @@ import numpy as np
 import io
 from PIL import Image
 from fabscan.scanner.interfaces.FSScanProcessor import FSScanProcessorCommand
-from fabscan.FSEvents import FSEvents
+from fabscan.FSEvents import FSEvents, FSEventManagerSingleton
 
 class FSStreamHandler(tornado.web.RequestHandler):
 
-    def initialize(self, scanprocessor):
+    def initialize(self, scanprocessor, eventmanager):
         self._logger = logging.getLogger(__name__)
         self.scanprocessor = scanprocessor.start()
+        self.eventmanager = eventmanager.get_instance()
+        self.stop_mjpeg = False
+
         self.types = {
             'laser':       FSScanProcessorCommand.GET_LASER_STREAM,
             'adjustment':  FSScanProcessorCommand.GET_ADJUSTMENT_STREAM,
             'texture':     FSScanProcessorCommand.GET_TEXTURE_STREAM,
             'calibration': FSScanProcessorCommand.GET_CALIBRATION_STREAM
         }
+        self.eventmanager.subscribe(FSEvents.ON_STOP_MJPEG_STREAM, self.on_mjpeg_stop)
 
+    def on_mjpeg_stop(self, mgr, event):
+        self.stop_mjpeg = True
 
     def getFrame(self, stream_type):
         try:
-            if self.scanprocessor.is_alive():
+            if not self.stop_mjpeg and self.scanprocessor.is_alive():
                 if stream_type == "laser":
                     img = self.scanprocessor.ask({FSEvents.COMMAND: FSScanProcessorCommand.GET_SETTINGS_STREAM})
                 else:
@@ -38,7 +44,7 @@ class FSStreamHandler(tornado.web.RequestHandler):
             return jpeg.tostring()
 
         except Exception as e:
-           self._logger.warning("Error while trying to trigger the scan processor: " + str(e))
+           self._logger.warning("Error while trying to trigger the scan processor: {0}".format(e))
 
 
     @tornado.web.asynchronous
@@ -49,6 +55,7 @@ class FSStreamHandler(tornado.web.RequestHandler):
         input: None
         :return: yields mjpeg stream with http header
         """
+        self._logger.debug("mjpeg stream started.")
         stream_type = self.get_argument('type', True)
         # Set http header fields
         self.set_header('Cache-Control',
@@ -57,17 +64,20 @@ class FSStreamHandler(tornado.web.RequestHandler):
         self.set_header('Content-Type', 'multipart/x-mixed-replace;boundary=--boundarydonotcross')
         self.set_header('Connection', 'close')
 
-        while True:
+        while not self.stop_mjpeg:
             try:
                 # Generating images for mjpeg stream and wraps them into http resp
                 img = self.getFrame(stream_type)
                 self.write("--boundarydonotcross\n")
                 self.write("Content-type: image/jpeg\r\n")
-                self.write("Content-length: %s\r\n\r\n" % len(img))
+                self.write("Content-length: {0}\r\n\r\n".format(img))
                 self.write(img)
                 yield tornado.gen.Task(self.flush)
             except Exception as e:
-                self._logger.warning('Stream canceled....' + str(e))
+                self._logger.warning("mjpeg stream stopped: {0}".format(e))
 
     def on_finish(self):
-        self.scanprocessor.stop()
+        time.sleep(2)
+        self.stop_mjpeg = False
+        self._logger.debug("Stream Handler Finished.")
+